@@ -10,20 +10,20 @@
   "author": "Alex Lobera <alex@leanjs.com>",
   "dependencies": {
     "@leanjs/core": "*",
+    "@leanjs/e2e-test-subjects-remote-react-1": "*",
     "@leanjs/next": "*",
     "@leanjs/react": "*",
-    "next": "^11.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "next": "^13.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@leanjs/webpack": "*",
     "@types/node": "^17.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "eslint": "8.15.0",
-    "eslint-config-next": "12.1.6",
-    "next-transpile-modules": "^9.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "eslint": "8.26.0",
+    "eslint-config-next": "^13.0.0",
     "typescript": "^4.8.3"
   }
 }
