@@ -13,8 +13,8 @@
     "build": "turbo run build"
   },
   "devDependencies": {
-    "@leanjs/cli": "^0.7.19",
-    "@leanjs/webpack": "^0.19.21",
+    "@leanjs/cli": "^0.7.21",
+    "@leanjs/webpack": "^0.20.2",
     "@leanjs/webpack-react": "^0.2.24",
     "@swc/core": "^1.2.142",
     "@swc/jest": "^0.2.17",
@@ -33,14 +33,14 @@
     "webpack": "^5.58.2"
   },
   "dependencies": {
-    "@leanjs/core": "0.20.0",
-    "@leanjs/react": "0.13.1",
+    "@leanjs/core": "0.21.1",
+    "@leanjs/react": "0.14.4",
     "@pixi/graphics-extras": "^6.5.4",
     "@reduxjs/toolkit": "^1.8.5",
     "history": "5.3.0",
     "pixi.js": "6.4.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-redux": "^8.0.4",
     "react-router-dom": "6.3.0"
   },
