@@ -19,16 +19,16 @@
     "@leanjs/react": "^0.14.4",
     "@testing-library/react": "^12.1.4",
     "@types/react": "^17.0.41",
-    "next": "12",
-    "react": "17",
-    "react-dom": "17"
+    "next": "13",
+    "react": "18",
+    "react-dom": "18"
   },
   "peerDependencies": {
     "@leanjs/core": "^0.13.4",
     "@leanjs/react": "^0.5.1",
-    "next": "12",
-    "react": "17",
-    "react-dom": "17"
+    "next": "13",
+    "react": "18",
+    "react-dom": "18"
   },
   "scripts": {
     "prebuild": "del-cli dist",
