@@ -1,1 +1,1 @@
-export { createRuntime } from "@leanjs/e2e-test-subjects-package-runtime-mocks";
+export { createRuntime } from "@leanjs/e2e-test-package-runtime-mocks";
