@@ -1,5 +1,5 @@
 import { createRuntimeBindings } from "@leanjs/react";
-import { createRuntime } from "@leanjs/e2e-test-subjects-package-runtime-shared";
+import { createRuntime } from "@leanjs/e2e-test-package-runtime-shared";
 
 export { createRuntime };
 
