@@ -1,13 +1,13 @@
+import React, { Suspense } from "react";
 import type { NextPage } from "next";
 import { Host } from "@leanjs/next";
+import ReactApp from "@leanjs/e2e-test-subjects-remote-react-1";
 
-const Home: NextPage = () => {
-  return (
-    <>
-      <h1>Nextjs Host</h1>
-      <Host app={{ packageName: "@leanjs/e2e-test-subjects-remote-react-1" }} />
-    </>
-  );
-};
+const Home: NextPage = () => (
+  <>
+    <h1>Nextjs Host</h1>
+    <Host app={ReactApp} />
+  </>
+);
 
 export default Home;
