@@ -1,15 +1,14 @@
+import React, { Suspense } from "react";
 import type { NextPage } from "next";
 import { Host } from "@leanjs/next";
 
-const ReactRedux: NextPage = () => {
-  return (
-    <>
-      <h1>React Redux page</h1>
-      <Host
-        app={{ packageName: "@leanjs/e2e-test-subjects-remote-react-redux" }}
-      />
-    </>
-  );
-};
+import App from "@leanjs/e2e-test-subjects-remote-react-redux";
+
+const ReactRedux: NextPage = () => (
+  <>
+    <h1>React Redux page</h1>
+    <Host app={App} />
+  </>
+);
 
 export default ReactRedux;
