@@ -10,6 +10,6 @@ export default defineNuxtConfig({
     },
   },
   build: {
-    transpile: ["@leanjs/e2e-test-subjects-package-runtime-shared"],
+    transpile: ["@leanjs/e2e-test-package-runtime-shared"],
   },
 });
