@@ -3,7 +3,7 @@
 describe("Nextjs shell: error", () => {
   it("displays an error message when the remote app fails", () => {
     cy.intercept("**/remoteEntry.js", { statusCode: 500 });
-    cy.visit("http://localhost:44440");
+    cy.visit("http://localhost:44440/react-error");
 
     cy.contains("Error: Failed to load script").should("be.visible");
   });
