@@ -12,27 +12,19 @@ describe("Nextjs shell: loading", () => {
     }).as("remoteEntry");
   });
 
-  it("displays a default loading component", () => {
-    cy.contains("...").should("not.exist");
-    cy.visit("http://localhost:44440");
-    cy.contains("Vue page").click();
-    cy.contains("...").should("be.visible");
-  });
-
   it("displays a custom loading component", () => {
     cy.visit("http://localhost:44440");
-    cy.contains("React sub page").click();
-    cy.contains("...").should("not.exist");
+    cy.contains("Vue page").click();
     cy.contains("Loading micro frontend").should("be.visible");
   });
 
   it("doesn't display a loading component the second time the remote app is displayed", () => {
     cy.visit("http://localhost:44440");
     cy.contains("Vue page").click();
-    cy.contains("...").should("be.visible");
+    cy.contains("Loading micro frontend").should("be.visible");
 
     cy.contains("React page").click();
     cy.contains("Vue page").click();
-    cy.contains("...").should("not.exist");
+    cy.contains("Loading micro frontend").should("not.exist");
   });
 });
